backend/requirements.txt
Flask==2.3.2
mysql-connector-python==9.0.0
gunicorn==21.2.0
flask-cors==4.0.0
